refactor: 모든 ChristmasPromotionPrecondition 의 구현 클래스
PROMOTION_NAME 의 타입을 PromotionName 열거형으로 수정합니다.
DiscountAmount 의 리팩터링에 따라 수정합니다.
1. apply 메소드가 Optional<DiscountInfo> 를 반환하도록 합니다.
2. calculateDiscountAmount 는 계산된 금액을 정수로 반환합니다.
